Выделяет логику отрисовки попапа в отдельный презентер
