+++ src/model/films-model.js
@@ -1,7 +1,7 @@
 import {generateFilm} from '../mock/film.js';
 
 export default class FilmsModel {
-  #films = Array.from({length: 23}, generateFilm);
+  #films = Array.from({length: 14}, generateFilm);
 
   get films() {
     return this.#films;
